pyproject.toml
[project]
name = "swp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.16.1",
    "anyio>=4.9.0",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.30.0",
    "authlib==1.3.0",
    "black>=25.1.0",
    "colorlog>=6.9.0",
    "fastapi[standard]>=0.112",
    "greenlet>=3.2.2",
    "jinja2>=3.1.6",
    "jose>=1.0.0",
    "jwt>=1.3.1",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "passlib>=1.7.4",
    "psycopg>=3.2.9",
    "pyjwt>=2.10.1",
    "pytest>=8.4.0",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.2.1",
    "python-dotenv>=1.1.0",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.41",
    "sqlmodel>=0.0.24",
]
